Document time parsing helpers
